refactor(csv): extract export helpers
